+++ src/main/java/com/miniconomy/commercial_bank_service/response/DebitOrderResponse.java
@@ -16,7 +16,7 @@ public class DebitOrderResponse
   private String creditAccountName;
   private String debitAccountName;
   private String debitOrderCreatedDate;
-  private BigDecimal debitOrderAmount;
+  private Long debitOrderAmount;
   private String debitOrderReceiverRef;
   private String debitOrderSenderRef;
   private boolean debitOrderDisabled;
